Refactor NetworkingImage
